+++ lucene/core/src/java/org/apache/lucene/search/similarities/LMDirichletSimilarity.java
@@ -114,10 +114,10 @@ public class LMDirichletSimilarity extends LMSimilarity {
   protected Explanation explain(
       BasicStats stats, int doc, Explanation freq, double docLen) {
     List<Explanation> subs = new ArrayList<>();
-    explain(subs, stats, doc, freq.getValue(), docLen);
+    explain(subs, stats, doc, freq.getValue().doubleValue(), docLen);
 
     return Explanation.match(
-        (float) score(stats, freq.getValue(), docLen),
+        (float) score(stats, freq.getValue().doubleValue(), docLen),
         "score(" + getClass().getSimpleName() + ", doc=" + doc + ", freq=" +
             freq.getValue() +"), computed as boost * " +
             "(term weight + document norm) from:",
